Move some helpers outside from the test module
