+++ tests/functional/testplan/runners/pools/test_pool_remote.py
@@ -4,46 +4,14 @@ import os
 import pytest
 import shutil
 import tempfile
-import subprocess
 
-import testplan
 from testplan.runners.pools import RemotePool
-from testplan.common.utils.remote import copy_cmd
+
 from .func_pool_base_tasks import schedule_tests_to_pool
 
+# tests/helpers are on the pythonpath
 from pytest_test_filters import skip_on_windows
-
-
-def mock_ssh(host, command):
-    """Avoid network connection."""
-    return ["/bin/sh", "-c", command]
-
-
-def strip_host(source, target, **kwargs):
-    """Avoid network connection."""
-    if ":" in source:
-        source = source.split(":")[1]
-    if ":" in target:
-        target = target.split(":")[1]
-    return copy_cmd(source, target)
-
-
-def copytree(src, dst):
-    """
-    We can't use shutil.copytree() with python 3.4.4 due to
-    https://bugs.python.org/issue21697 so use rsync instead.
-    """
-    subprocess.check_call(
-        [
-            "rsync",
-            "-rL",
-            "--exclude=.git",
-            "--exclude=*.pyc",
-            "--exclude=*__pycache__*",
-            src,
-            dst,
-        ]
-    )
+from remote_utils import mock_ssh, strip_host, copytree
 
 
 def setup_workspace():
